requirements.txt
requests~=2.32.3
tqdm~=4.66.4
beautifulsoup4~=4.12.3
bs4~=0.0.1
python-dotenv~=1.0.1